Added 'htmlcompress' => true, at renderer in config/settings.php Added module/bootstrap.js Added validation in module/create-folder.js Added backslash recognition in module/two-factor.js Added validation and placeholder in module/update-file.js Added validation and placeholder in module/update-folder.js Added textarea and placeholder style in module/_modal.scss Added redirect in public/.htaccess Added sql/upload_limit-records.sql Added note handling at updateFileAction() in Controller/ApiController.php Added note handling at getFileAction() in Controller/ApiController.php Added return type at all setter in MappedSuperclass/Base.php Added return type at all setter in MappedSuperclass/BaseJoin.php Added MappedSuperclass/BaseUuid.php Added $settings['renderer']['htmlcompress'] in src/dependencies.php Added note textarea in partial/modal-update-file.html.twig Added partial/modal-create-folder.html.twig in folder/show.html.twig Added platform in composer.json Added doctrine/annotations 1.13 Added ramsey/uuid-doctrine 1.8 Added symfony/cache 6.0 Added datatables.net-bs5 1.13.1 Added sass 1.56.1
Changed bools to lowercase and dbname in config/additional-settings.dist.php
Changed bools to lowercase in config/settings.php
Changed proxy in app/gulpfile.dist.json
Changed BS4 to BS5 syntax in module/cookieconsent.js
Changed access button template in module/create-folder.js
Changed access button and folder button template in module/update-file.js
Changed access button template in module/update-folder.js
Changed BS4 to BS5 syntax in js/scripts.js
Changed BS4 to BS5 syntax in lib/bootstrap.scss
Changed FA5 to FA6 syntax in lib/fontawesome.scss
Changed button background color in module/_cookieconsent.scss
Changed button group input style in module/_general.scss
Changed hierarchy in module/_switch.scss
Changed font family in module/_timeline.scss
Changed BS4 to BS5 syntax in plugin/datatables.bootstrap.scss
Changed various in scss/_variables.scss
Changed locale/de-DE.php
Changed locale/en-US.php
Changed nickname in public/.htaccess
Changed imhhfs_* to fs_* in sql/*.sql
Changed Composer/Setup.php
Changed Controller/ApiController.php
Changed imhhfs_* to fs_* in Entity/*.php
Changed extends Base to extends BaseUuid in Entity/File.php
Changed extends Base to extends BaseUuid in Entity/Folder.php
Changed extends Base to extends BaseUuid in Entity/RecoveryCode.php
Changed localeQualityAsc() for php 8.0 in Utility/LanguageUtility.php
Changed templates to template
Changed templates/layouts to template/layout
Changed templates/partials to template/partial
Changed BS4 to BS5 syntax in all twig files
Changed .gitignore
Changed php:7.4.2 to php:8.0 in Dockerfile
Changed README.md
Changed deploy.sh
Changed ports and database name in docker-compose.yml
Changed nickname and datatables.net-bs4 to datatables.net-bs5 in gulpfile.js
Changed doctrine/orm 2.5 to 2.13
Changed monolog/monolog 1.17 to 1.27
Changed slim/slim 3.1 to 3.12
Changed slim/twig-view 2.3 to 2.5
Changed symfony/console 3.0|4.0 to 6.0
Changed @fortawesome/fontawesome-free 5.15.2 to 6.2.1
Changed bootstrap 4.6.0 to 5.2.3
Changed browser-sync 2.26.14 to 2.27.10
Changed datatables.net 1.10.23 to 1.13.1
Changed del 6.0.0 to 6.1.1
Changed gulp-autoprefixer 7.0.1 to 8.0.0
Changed gulp-sass 4.1.0 to 5.1.0
Changed gulp-uglify-es 2.0.0 to 3.0.0
Changed jquery 3.5.1 to 3.6.1

Fixed typo in routes/xx-XX.php
Fixed undefined index in Controller/ErrorController.php
Fixed return type at getFile() and setFile() in Entity/File.php
Fixed return type at getCurrentUser() and encryptPassword() in Utility/GeneralUtility.php

Removed $id in Entity/File.php
Removed $id in Entity/Folder.php
Removed $id in Entity/RecoveryCode.php
Removed unused findAll() in Repository/UserRepository.php
Removed composer.lock
Removed package-lock.json
Removed datatables.net-bs4

+++ src/Controller/ApiController.php
@@ -6,13 +6,14 @@ use App\Entity\File;
 use App\Entity\FileFolderJoin;
 use App\Entity\Folder;
 use App\Entity\User;
+use App\Utility\GeneralUtility;
 use App\Utility\LanguageUtility;
 
 /**
  * ApiController is used for request via ajax
  */
 class ApiController extends BaseController {
-    
+
     /**
      * Create folder Action
      * 
@@ -23,13 +24,16 @@ class ApiController extends BaseController {
      */
     public function createFolderAction($request, $response, $args) {
         $user = $this->em->getRepository('App\Entity\User')->findOneBy(['id' => $this->currentUser]);
-        
+
         // if request is xhr and user is logged in
+        // AND folder name is greater than 2 chars
+        // AND access is greater than 0
         if ($request->isXhr() && $user instanceof User 
-                && strlen($request->getParam('name')) > 2 && (int)$request->getParam('access')) {
+                && strlen($request->getParam('name')) > 2 
+                && (int)$request->getParam('access')) {
             $host = ($request->getServerParam('SERVER_PORT') == '80' ? 'http' : 'https') . '://' . $request->getServerParam('HTTP_HOST');
             $access = $this->em->getRepository('App\Entity\Access')->findOneBy(['id' => (int)$request->getParam('access')]);
-            
+
             if ($access instanceof Access) {
                 $folder = new Folder();
                 $folder->setName($request->getParam('name'))
@@ -54,10 +58,10 @@ class ApiController extends BaseController {
                 ]);
             }
         }
-        
+
         return json_encode(['result' => FALSE]);
     }
-    
+
     /**
      * Get file Action
      * 
@@ -69,24 +73,31 @@ class ApiController extends BaseController {
     public function getFileAction($request, $response, $args) {
         $user = $this->em->getRepository('App\Entity\User')->findOneBy(['id' => $this->currentUser]);
         $file = $this->em->getRepository('App\Entity\File')->findOneBy(['id' => $args['uuid']]);
-        
+
         // if request is xhr and file exists
         if ($request->isXhr() && $file instanceof File) {
             $host = ($request->getServerParam('SERVER_PORT') == '80' ? 'http' : 'https') . '://' . $request->getServerParam('HTTP_HOST');
-            
-            // if files is public or shareable 
+
+            // if file is public or shareable 
             // OR logged in user is owner of file 
             // OR logged in user has role 'superasmin'
             if(($file->getAccessId() === 2 || $file->getAccessId() === 3) 
                     || ($user instanceof User && $user->getFiles()->contains($file)) 
                     || $this->currentRole === 'superadmin') {
                 $access = $file->getAccess();
-                
+
+                $note = '';
+                if ($file->getFile() instanceof File 
+                        && is_readable($this->settings['upload']['path'] . $file->getFile()->getHashName() . $file->getFile()->getExtension()->getName())) {
+                    $note = file_get_contents($this->settings['upload']['path'] . $file->getFile()->getHashName() . $file->getFile()->getExtension()->getName());
+                }
+
                 return json_encode([
                     'result' => TRUE,
                     'id' => $file->getId(),
                     'name' => $file->getName(),
                     'size' => $file->getSize(),
+                    'note' => $note,
                     'access' => $file->getAccessId(),
                     'access_icon' => $access->getIcon(),
                     'access_button' => $access->getButton(),
@@ -100,10 +111,10 @@ class ApiController extends BaseController {
                 ]);
             }
         }
-        
+
         return json_encode(['result' => FALSE]);
     }
-    
+
     /**
      * Get folder Action
      * 
@@ -115,19 +126,19 @@ class ApiController extends BaseController {
     public function getFolderAction($request, $response, $args) {
         $user = $this->em->getRepository('App\Entity\User')->findOneBy(['id' => $this->currentUser]);
         $folder = $this->em->getRepository('App\Entity\Folder')->findOneBy(['id' => $args['uuid']]);
-        
+
         // if request is xhr and file exists
         if ($request->isXhr() && $folder instanceof Folder) {
             $host = ($request->getServerParam('SERVER_PORT') == '80' ? 'http' : 'https') . '://' . $request->getServerParam('HTTP_HOST');
-            
-            // if files is public or shareable 
+
+            // if folder is public or shareable 
             // OR logged in user is owner of file 
             // OR logged in user has role 'superasmin'
             if(($folder->getAccessId() === 2 || $folder->getAccessId() === 3) 
                     || ($user instanceof User && $user->getFolders()->contains($folder)) 
                     || $this->currentRole === 'superadmin') {
                 $access = $folder->getAccess();
-                
+
                 return json_encode([
                     'result' => TRUE,
                     'id' => $folder->getId(),
@@ -143,10 +154,10 @@ class ApiController extends BaseController {
                 ]);
             }
         }
-        
+
         return json_encode(['result' => FALSE]);
     }
-    
+
     /**
      * Get access list Action
      * 
@@ -157,7 +168,7 @@ class ApiController extends BaseController {
      */
     public function getAccessListAction($request, $response, $args) {
         $accessArray = $this->em->getRepository('App\Entity\Access')->findAllArray();
-        
+
         // if request is xhr and $accessArray not empty
         if ($request->isXhr() && count($accessArray)) {
             return json_encode([
@@ -165,10 +176,10 @@ class ApiController extends BaseController {
                 'access_list' => $accessArray,
             ]);
         }
-        
+
         return json_encode(['result' => FALSE]);
     }
-    
+
     /**
      * Update file update Action
      * 
@@ -180,35 +191,68 @@ class ApiController extends BaseController {
     public function updateFileAction($request, $response, $args) {
         $user = $this->em->getRepository('App\Entity\User')->findOneBy(['id' => $this->currentUser]);
         $file = $this->em->getRepository('App\Entity\File')->findOneBy(['id' => $args['uuid']]);
-        
-        // if logged in user is owner of file 
+
+        // if logged in user is owner of file
+        // AND file name is greater than 2 chars
         // OR logged in user has role 'superasmin'
         if ($request->isXhr() && $file instanceof File 
                 && ($user instanceof User && $user->getFiles()->contains($file)) 
+                && strlen($request->getParam('name')) > 2 
                 || $this->currentRole === 'superadmin') {
             $host = ($request->getServerParam('SERVER_PORT') == '80' ? 'http' : 'https') . '://' . $request->getServerParam('HTTP_HOST');
             $file->setName($request->getParam('name'));
+            $note = $request->getParam('note');
             $accessParam = $this->em->getRepository('App\Entity\Access')->findOneBy(['id' => (int)$request->getParam('access')]);
-            
+
             if ($accessParam instanceof Access) {
                 $file->setAccess($accessParam);
             }
-            
+
+            // if file has a note
+            if ($file->getFile() instanceof File) {
+                $fileNote = $file->getFile();
+                $fileNote->setSize(strlen($note));
+                file_put_contents($this->settings['upload']['path'] . $fileNote->getHashName() . $fileNote->getExtension()->getName(), $note);
+            } else if (!($file->getFile() instanceof File) && strlen(trim($note))) {
+                $access = $this->em->getRepository('App\Entity\Access')->findOneBy(['id' => 1]);
+                $noteExtension = $this->em->getRepository('App\Entity\FileExtension')->findOneBy(['name' => '.txt']);
+
+                do {
+                    $noteFileName = 'note-' . GeneralUtility::generateCode(10) . '.txt';
+                    $noteHashName = GeneralUtility::generateCode(10) . substr(md5($noteFileName), 0, 10);
+                } while (file_exists($this->settings['upload']['path'] . $noteHashName . $noteExtension->getName()));
+
+                file_put_contents($this->settings['upload']['path'] . $noteHashName . $noteExtension->getName(), $note);
+
+                $fileNote = new File();
+                $fileNote->setName($noteFileName)
+                    ->setHashName($noteHashName)
+                    ->setMimeType('text/plain')
+                    ->setSize(strlen($note))
+                    ->setExtension($noteExtension)
+                    ->setAccess($access)
+                    ->setFileIncluded(true)
+                    ->setUser($user);
+
+                $file->setFile($fileNote);
+                $this->em->persist($fileNote);
+            }
+
             $access = $file->getAccess();
             $folders = $request->getParam('folders');
-            
+
             // remove all folder joins
             $file->getFolderJoins()->forAll(function($key, $entity) {
                 $this->em->remove($entity); 
                 return true;
             });
             $this->em->flush();
-            
+
             // set folder joins
             if (is_array($folders) && count($folders) > 0) {
                 foreach ($folders as $uuid) {
                     $folder = $this->em->getRepository('App\Entity\Folder')->findOneBy(['id' => $uuid]);
-                    
+
                     // if $folder is Folder
                     if ($folder instanceof Folder) {
                         $fileFolderJoin = new FileFolderJoin();
@@ -224,6 +268,7 @@ class ApiController extends BaseController {
                 'id' => $file->getId(),
                 'name' => $file->getName(),
                 'size' => $file->getSize(),
+                'note' => $note,
                 'access' => $file->getAccessId(),
                 'access_icon' => $access->getIcon(),
                 'access_button' => $access->getButton(),
@@ -236,10 +281,10 @@ class ApiController extends BaseController {
                 'user_link' => $host . $this->router->pathFor('user-show-' . LanguageUtility::getLocale(), ['name' => $file->getUser()->getName()]),
             ]);
         }
-        
+
         return json_encode(['result' => FALSE]);
     }
-    
+
     /**
      * Update folder Action
      * 
@@ -251,23 +296,25 @@ class ApiController extends BaseController {
     public function updateFolderAction($request, $response, $args) {
         $user = $this->em->getRepository('App\Entity\User')->findOneBy(['id' => $this->currentUser]);
         $folder = $this->em->getRepository('App\Entity\Folder')->findOneBy(['id' => $args['uuid']]);
-        
+
         // if logged in user is owner of file 
+        // AND folder name is greater than 2 chars
         // OR logged in user has role 'superasmin'
         if ($request->isXhr() && $folder instanceof Folder 
                 && ($user instanceof User && $user->getFolders()->contains($folder)) 
+                && strlen($request->getParam('name')) > 2 
                 || $this->currentRole === 'superadmin') {
             $host = ($request->getServerParam('SERVER_PORT') == '80' ? 'http' : 'https') . '://' . $request->getServerParam('HTTP_HOST');
             $folder->setName($request->getParam('name'));
             $accessParam = $this->em->getRepository('App\Entity\Access')->findOneBy(['id' => (int)$request->getParam('access')]);
-            
+
             if ($accessParam instanceof Access) {
                 $folder->setAccess($accessParam);
             }
-            
+
             $access = $folder->getAccess();
             $this->em->flush();
-            
+
             return json_encode([
                 'result' => TRUE,
                 'id' => $folder->getId(),
@@ -282,7 +329,7 @@ class ApiController extends BaseController {
                 'user_link' => $host . $this->router->pathFor('user-show-' . LanguageUtility::getLocale(), ['name' => $folder->getUser()->getName()]),
             ]);
         }
-        
+
         return json_encode(['result' => FALSE]);
     }
 }
